# Feat: Improve empty group edge case performance