```
goos: linux
goarch: amd64
pkg: github.com/samber/slog-common
cpu: AMD Ryzen 9 5950X 16-Core Processor            
                      │ baseline.txt │               pr.txt                │
                      │    sec/op    │   sec/op     vs base                │
AppendAttrsToGroup-32    144.7n ± 3%   110.5n ± 3%  -23.61% (p=0.000 n=10)

                      │ baseline.txt │               pr.txt               │
                      │     B/op     │    B/op     vs base                │
AppendAttrsToGroup-32     208.0 ± 0%   160.0 ± 0%  -23.08% (p=0.000 n=10)

                      │ baseline.txt │               pr.txt               │
                      │  allocs/op   │ allocs/op   vs base                │
AppendAttrsToGroup-32     3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=10)
```